backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyMySQL
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
python-decouple
httpx